Add write support for >64K intel hex files
git-svn-id: https://svn.code.sf.net/p/ec2drv/code/ec2drv/trunk@182 07914c0c-2d3e-0410-adf3-90c1a7ff5250

+++ src/ec2tools/ihex.h
@@ -6,7 +6,9 @@
 void ihex_load_file( char *filename, uint8_t *memory,
 					 uint32_t *start, uint32_t *end  );
 
-void ihex_save_file( char *filename, char *memory, uint16_t start_addr, uint16_t len );
-void ihex_hexout( FILE *fhex, int byte, int memory_location, int end );
+void ihex_save_file( char *filename, char *memory,
+					 uint32_t start_addr, uint32_t len );
+void ihex_hexout( FILE *fhex, int byte,
+				  uint32_t memory_location, uint32_t end );
 
 #endif
